requirements.txt
boto3>=1.26.0
pytest>=7.3.1
pytest-mock>=3.10.0
pyyaml>=6.0.1